requirements.txt
google-genai
python-dotenv
streamlit
PyGithub
